TEST: add util/opensslenv.sh and use it
util/opensslenv.sh is a script a little in between util/shlib_wrap.sh
and util/opensslwrap.sh, and is mostly useful to create and display
commands in the test that are easy to copy and paste without having to
think about extra environment variables.

This is only useful in a POSIX shell, the other environments we
support still have different mechanisms.